Health Comparison Spreadsheet export
Because
The health insurance comparison show view only provides a very simple
view of the data, it doesn't give sufficient information about the
actual benefit coverage.

Allowing an excel export means the full benefit limit information can
be provided in a spreadsheet with navigation between the different
benefit categories handled by a spreadsheet tab for each category

Each benefit cell is colour-coded according to coverage. Green when the
benefit is paid in full. Orange when it is capped or red when it isn't
covered by that plan in the comparison.

This commit:

* Add caxslx, caxlsx_rails and axlsx_styler gems
* Add excel logo svg
* Add overall_sum_assured method to HealthInsurancePolicy model
* Add xlsx response format to health plan comparisons controller show method
* Add benefit_limit method to NullProductModuleMedicalBenefit model
* Add export link to health comparison show page
* Add xlsx health plan comparison show view page